requirements.txt
flask>=2.3.0
matplotlib==3.4.3
seaborn==0.11.2
python-dotenv
flask-cors
gunicorn
